refactor(mojo): extract analysis/spread/memo/props domains
Round out the Mojo adapter's go-template-style decomposition (issue #2018
track D) with the remaining domain modules, each behind a narrow context
seam where it needs adapter callbacks. Behaviour unchanged — conformance
byte-identical (464 pass / 0 fail).

  analysis/component-tree.ts  hasClientInteractivity,
                              collectImportedLoopChildComponentErrors
                              (returns diagnostics; adapter owns `errors`)
  spread/spread-codegen.ts    conditional-spread / object-literal → Perl
                              hashref lowering (MojoSpreadContext seam)
  memo/seed.ts                in-template derived-memo / context seeding
                              (MojoMemoContext seam)
  props/prop-classes.ts       per-compile prop classification sets

`type/` is intentionally absent: unlike the Go adapter, the Mojo
template target is dynamically typed and emits no struct/type codegen.
